feat: Adding logging and tracing via aspire dashboard
When running the stack it will configure Open Telemetry logging for the api and echo-agent services, and setup an aspire dashboard as the OTLP exporter destination. This allows visualisation of the logs, as well as traces across service boundaries (ie - viewing the request flow from MCP client to MCP server). I've also included basic metric counting (connections and messages) to illustrate how we can get that level of insights.