feat: add .NET tool scaffolding (#16)
* chore: Creating a blank project for .NET agent
* chore: Moving to the right location
* chore: Running .NET server in a container
* feat: Basic .NET MCP server
Created a MCP Server on SSE (using a custom SSE endpoint implementation as the SDK doesn't provide one yet) that exposes an 'echo' tool.

Updated the docker-compose to support this properly, also aligned the environment variables across services.
Integrated a call to the server from the API client, and refactored that a bit to make the code a little easier to read